Update super_csv_splitter.sh
Added two new IF statements that check if the user input is either empty, or a full stop (period), in which case the current working directory will be used in the getFilePath() function.

This enables you to run the script as an alias from within your current working directory, which might make life easier for you.